+++ midealocal/devices/b1/__init__.py
@@ -81,7 +81,7 @@ class MideaB1Device(MideaDevice):
                 new_status[str(status)] = self._attributes[status]
         return new_status
 
-    def set_attribute(self, attr: str, value: bool | int | str) -> None:
+    def set_attribute(self, attr: str, value: bool | float | str) -> None:
         """Midea B1 device set attribute."""
