+++ midealocal/devices/b6/__init__.py
@@ -98,7 +98,7 @@ class MideaB6Device(MideaDevice):
                 new_status[str(status)] = self._attributes[status]
         return new_status
 
-    def set_attribute(self, attr: str, value: bool | int | str) -> None:
+    def set_attribute(self, attr: str, value: bool | float | str) -> None:
         """Midea B6 device set attribute."""
         message = None
         if attr == DeviceAttributes.fan_speed:
