+++ midealocal/devices/cc/__init__.py
@@ -219,7 +219,7 @@ class MideaCCDevice(MideaDevice):
             message.mode = mode
         self.build_send(message)
 
-    def _set_attribute_fe(self, attr: str, value: bool | int | str) -> None:
+    def _set_attribute_fe(self, attr: str, value: bool | float | str) -> None:
         """Set an attribute on a 0xFE VRF panel via key-value control frames."""
         if attr == DeviceAttributes.power:
             self._send_fe_control([(CCControlId.POWER, 1 if value else 0)])
@@ -244,7 +244,7 @@ class MideaCCDevice(MideaDevice):
             self._send_fe_control([(CCControlId.SWING, 0x06 if value else 0x00)])
         # other attributes are not supported by the 0xFE control protocol
 
-    def set_attribute(self, attr: str, value: bool | int | str) -> None:
+    def set_attribute(self, attr: str, value: bool | float | str) -> None:
         """Midea CC device set attribute."""
         if self._is_fe_format:
             self._set_attribute_fe(attr, value)
